separating platforms even more, TrueStudio project added (PART 2)

+++ eez/psu/sd_card.h
@@ -19,7 +19,7 @@
 #pragma once
 
 #if defined(EEZ_PLATFORM_SIMULATOR)
-#include "eez/psu/platform/simulator/arduino/SdFat.h"
+#include "eez/psu/platform/simulator/SdFat.h"
 #endif
 
 #if defined(EEZ_PLATFORM_ARDUINO_DUE)
@@ -48,7 +48,7 @@ bool makeParentDir(const char *filePath);
 bool exists(const char *dirPath, int *err);
 bool catalog(const char *dirPath, void *param, void (*callback)(void *param, const char *name, const char *type, size_t size), int *err);
 bool catalogLength(const char *dirPath, size_t *length, int *err);
-bool upload(const char *filePath, void *param, void (*callback)(void *param, const void *buffer, size_t size), int *err);
+bool upload(const char *filePath, void *param, void (*callback)(void *param, const void *buffer, int size), int *err);
 bool download(const char *filePath, bool truncate, const void *buffer, size_t size, int *err);
 bool moveFile(const char *sourcePath, const char *destinationPath, int *err);
 bool copyFile(const char *sourcePath, const char *destinationPath, int *err);
